refactor: [#100] rename module to dvc_diff_media_parser
