+++ src/nautilus_librarian/typer/commands/workflows/actions/check_images_changes.py
@@ -1,4 +1,4 @@
-from nautilus_librarian.domain.dvc_diff_parser import (
+from nautilus_librarian.domain.dvc_diff_media_parser import (
     extract_all_changed_files_from_dvc_diff,
 )
 from nautilus_librarian.typer.commands.workflows.actions.action_result import (
