+++ src/nautilus_librarian/typer/commands/workflows/actions/validate_images_dimensions_action.py
@@ -1,4 +1,4 @@
-from nautilus_librarian.domain.dvc_diff_parser import (
+from nautilus_librarian.domain.dvc_diff_media_parser import (
     extract_added_and_modified_files_from_dvc_diff,
 )
 from nautilus_librarian.mods.dvc.domain.utils import (
